requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
django-ranged-response==0.2.0
django-uuslug==1.2.0
Pillow==8.2.0
psycopg2-binary==2.8.6
python-slugify==4.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
